// Zybach.Web/src/app/shared/generated/model/sensor-summary-dto.ts
/**
 * Zybach
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class SensorSummaryDto { 
    WellRegistrationID?: string;
    SensorID?: number;
    SensorName?: string;
    SensorTypeID?: number;
    SensorType?: string;
    IsActive?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}